+++ web/packages/common/src/components/KeyValueGrid/index.tsx
@@ -4,17 +4,17 @@
 import { KVPair } from '@nemo/common/src/components/KVPair';
 import type { FC, ReactNode } from 'react';
 
-export interface KeyValueGridItem {
+interface KeyValueGridItem {
   key: string;
   label: string;
   value: ReactNode;
   wrapValue?: boolean;
 }
 
+// Booleans are excluded: React renders `true`/`false` as nothing, so they would
+// otherwise produce a labelled row with empty content.
 const isPresentValue = (value: ReactNode): boolean =>
-  value !== null && value !== undefined && value !== '';
-
-const KEY_VALUE_GRID_CLASS = 'flex flex-wrap min-w-0 gap-x-density-2xl gap-y-density-lg';
+  value !== null && value !== undefined && value !== '' && typeof value !== 'boolean';
 
 /**
  * Key/value fields laid out in a wrapping flow: each pair takes only the width
@@ -28,7 +28,7 @@ export const KeyValueGrid: FC<{ items: readonly KeyValueGridItem[] }> = ({ items
     return null;
   }
   return (
-    <div className={KEY_VALUE_GRID_CLASS}>
+    <div className="flex flex-wrap min-w-0 gap-x-density-2xl gap-y-density-lg">
       {visible.map((item) => (
         <KVPair
           key={item.key}
