+++ web/packages/studio/src/components/IntakeDetail/SpanTemplates/EvaluatorSpanContent.tsx
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { Badge } from '@nvidia/foundations-react-core';
 import {
   asBoolean,
   asNumber,
@@ -31,13 +32,23 @@ export const EvaluatorSpanContent: FC<SpanTemplateContentProps> = ({ span }) => 
   const testsTotal = asNumber(attributes['evaluator.tests_total']);
 
   const fields: TemplateField[] = [
+    // Score leads the kind fields so it renders immediately after Status, shown
+    // as a badge to set the headline verdict apart from the metadata that follows.
+    {
+      label: 'Score',
+      value:
+        score === undefined ? undefined : (
+          <Badge color="gray" kind="solid">
+            {score.toFixed(3)}
+          </Badge>
+        ),
+    },
     // The span name (e.g. "judge-answer") conveys what the evaluator does; the
     // row header elevates the evaluator's own name, so surface this here.
     { label: 'Name', value: span.name ?? undefined },
     { label: 'Evaluator', value: name },
     { label: 'Type', value: kind },
     { label: 'Result', value: passed === undefined ? undefined : passed ? 'Pass' : 'Fail' },
-    { label: 'Score', value: score?.toFixed(3) },
     {
       label: 'Tests',
       value: testsTotal === undefined ? undefined : `${testsPassed ?? 0} / ${testsTotal}`,
