+++ web/packages/studio/src/components/IntakeDetail/SpanTemplates/RerankerSpanContent.tsx
@@ -1,8 +1,13 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { IntakeAccordion } from '@nemo/common/src/components/IntakeAccordion';
-import { Stack, Text } from '@nvidia/foundations-react-core';
+// Exports the kind body component alongside its accordion-section builder; the
+// pair is small and co-located, so fast refresh's component-only rule is waived.
+/* eslint-disable react-refresh/only-export-components */
+
+import type { IntakeAccordionItem } from '@nemo/common/src/components/IntakeAccordion';
+import type { Span } from '@nemo/sdk/generated/platform/schema';
+import { Text } from '@nvidia/foundations-react-core';
 import {
   asNumber,
   extractRerankedDocuments,
@@ -20,45 +25,42 @@ const DOCUMENTS_SECTION = 'reranker-documents';
 
 /**
  * RERANKER body. The rerank model/top-N float side by side as key values; the
- * reranked document scores (`reranker.documents.*`) are a collapsible section,
- * open by default.
+ * reranked document scores are surfaced as their own accordion section (see
+ * `rerankerCustomSections`) so the Annotations section can lead the group.
  */
 export const RerankerSpanContent: FC<SpanTemplateContentProps> = ({ span }) => {
   const attributes = parseRawAttributes(span.raw_attributes);
-  const documents = extractRerankedDocuments(span);
   const topN = asNumber(attributes['reranker.top_n']);
-  const documentsLabel = documents.length
-    ? `Ranked documents (${documents.length})`
-    : 'Ranked documents';
 
   const fields: TemplateField[] = [
     { label: 'Model', value: span.model ?? undefined },
     { label: 'Top N', value: topN?.toLocaleString() },
   ];
 
-  return (
-    <Stack gap="density-xl" className="min-w-0">
-      <TemplateKeyValues span={span} fields={fields} />
-      <IntakeAccordion
-        variant="section"
-        defaultValue={[DOCUMENTS_SECTION]}
-        items={[
-          {
-            value: DOCUMENTS_SECTION,
-            slotLabel: (
-              <Text kind="body/semibold/sm" className="min-w-0">
-                {documentsLabel}
-              </Text>
-            ),
-            slotContent: (
-              <RankedDocumentList
-                documents={documents}
-                emptyMessage="No reranked documents were captured for this span."
-              />
-            ),
-          },
-        ]}
-      />
-    </Stack>
-  );
+  return <TemplateKeyValues span={span} fields={fields} />;
+};
+
+/** Reranker accordion sections: the reranked document scores (`reranker.documents.*`). */
+export const rerankerCustomSections = (span: Span): IntakeAccordionItem[] => {
+  const documents = extractRerankedDocuments(span);
+  const documentsLabel = documents.length
+    ? `Ranked documents (${documents.length})`
+    : 'Ranked documents';
+
+  return [
+    {
+      value: DOCUMENTS_SECTION,
+      slotLabel: (
+        <Text kind="body/semibold/sm" className="min-w-0">
+          {documentsLabel}
+        </Text>
+      ),
+      slotContent: (
+        <RankedDocumentList
+          documents={documents}
+          emptyMessage="No reranked documents were captured for this span."
+        />
+      ),
+    },
+  ];
 };
