+++ web/packages/studio/src/components/IntakeDetail/SpanTemplates/RetrieverSpanTemplate.ts
@@ -1,12 +1,16 @@
 // SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { RetrieverSpanContent } from '@studio/components/IntakeDetail/SpanTemplates/RetrieverSpanContent';
+import {
+  RetrieverSpanContent,
+  retrieverCustomSections,
+} from '@studio/components/IntakeDetail/SpanTemplates/RetrieverSpanContent';
 import type { SpanTemplate } from '@studio/components/IntakeDetail/SpanTemplates/types';
 
-/** Retriever: query + ranked documents live in the kind section; no output, no usage. */
+/** Retriever: query + ranked documents are accordion sections; no output, no usage. */
 export const retrieverSpanTemplate: SpanTemplate = {
   Content: RetrieverSpanContent,
   sections: ['kind', 'metadata', 'annotations'],
   attributeNamespaces: ['retrieval'],
+  customSections: retrieverCustomSections,
 };
